Moved DamageCalculation script to stats folder. Created simple enemy to test attacks. Added Audio Manager to my scene. Edited arrowscript for testing. Made debug message for controller change a warning. Added RangedMultiple Attacks to BFA. Fixed issue where I had not passed the Projectile prefabs to the new created object for BFA

+++ Assets/Scripts/Player/Stats/PlayerStat.cs
@@ -26,7 +26,7 @@ public class PlayerStat : ScriptableObject
 
 
 	[Header("Range")]
-	public float MultipleNumber;
+	public int MultipleNumber;
 	public float MultipleRange;
 	public float AreaOfEffect;
 
@@ -78,7 +78,7 @@ public class PlayerStat : ScriptableObject
 		this.MeleeRange = 0f;
 		this.MeleeReach = 0f;
 
-		this.MultipleNumber = 0f;
+		this.MultipleNumber = 0;
 		this.MultipleRange = 0f;
 		this.AreaOfEffect = 0f;
